+++ CHANGELOG.md
@@ -23,7 +23,7 @@
 * Fix missing entry for Trigger in TektonTriggersResourceMappingProvider
 * Fix #3047: NPE when getting version when there is no build date
 * Fix #3024: stopAllRegisteredInformers will not call startWatcher
-* Fix #3067: Added a patch operation with an explicit base
+* Fix #3067: Added a patch(patchcontet, item) operation to be more explicit about patching and diffing behavior
 
 #### Improvements
 * Fix #2788: Support FIPS mode in kubernetes-client with BouncyCastleFipsProvider
@@ -35,7 +35,7 @@
 * Fix #3061: Removed the deltafifo from the informer logic
 * Use apiGroupName in generated package for OpenShiftClient Handler/OperationsImpl classes
 * Fix #3089: Allowing patch/edit to infer context from the item
-* Fix #3066: Added an applyStatus method
+* Fix #3066: Added replaceStatus (PUT), editStatus (JSON PATCH), and patchStatus (JSON MERGE PATCH) methods to support non-locking status updates
 
 #### Dependency Upgrade
 * Fix #2979: Update Kubernetes Model to v1.21.0